% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{initLogger}
\alias{initLogger}
\title{Init logger}
\usage{
initLogger(threshold = "DEBUG")
}
\arguments{
\item{threshold}{The logging threshold level. Messages with a lower priority
level will be discarded. See \code{\link[log4r]{loglevel}}.}
}
\description{
Init logger
}
\seealso{
\link[log4r]{logger}
}
