% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{inputStateContainer}
\alias{inputStateContainer}
\title{Container to allow set state for inputs}
\usage{
inputStateContainer(id, inputUI)
}
\arguments{
\item{id}{string, id of the input}

\item{inputUI}{string/shiny.tag, input content}
}
\value{
shiny.tag of the panel
}
\description{
Container to allow set state for inputs
}
