% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logError}
\alias{logError}
\title{Log error message}
\usage{
logError(...)
}
\arguments{
\item{...}{One or more items to be written to the log at the corresponding
priority level.}
}
\description{
Log error message
}
\seealso{
\link[log4r]{error}
}
