% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logInfo}
\alias{logInfo}
\title{Log info message}
\usage{
logInfo(...)
}
\arguments{
\item{...}{One or more items to be written to the log at the corresponding
priority level.}
}
\description{
Log info message
}
\seealso{
\link[log4r]{info}
}
