% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logWarn}
\alias{logWarn}
\title{Log warning message}
\usage{
logWarn(...)
}
\arguments{
\item{...}{One or more items to be written to the log at the corresponding
priority level.}
}
\description{
Log warning message
}
\seealso{
\link[log4r]{warn}
}
