% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{panel}
\alias{panel}
\title{Bootstrap panel}
\usage{
panel(body, head = NULL, footer = NULL, collapsible = FALSE)
}
\arguments{
\item{body}{string/shiny.tag, panel body content}

\item{head}{string/shiny.tag, panel header content}

\item{footer}{string/shiny.tag, panel footer content}

\item{collapsible}{flag to turn on/off collapsing functionality}
}
\value{
shiny.tag of the panel
}
\description{
Function that creates the UI of the bootstrap panel
}
\seealso{
\href{https://getbootstrap.com/docs/3.3/components/#panels}{Bootstrap panels}
}
