% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{setInputState}
\alias{setInputState}
\title{Set input state}
\usage{
setInputState(session, id, state = NULL)
}
\arguments{
\item{session}{shiny session object}

\item{id}{string, id of the input}

\item{state}{string/NULL, "success"|"warning"|"error"|NULL}
}
\description{
Set input state
}
